Rename package to internal to make it clear its for netty use only.
Motivation:

We should put everything in an internal package to make it clear its for netty usage only and so we are free to break it at any time.

Modifications:

Move everything to an internal package.

Result:

Ensure peple are aware its an internal project.
