+++ openssl-dynamic/src/main/java/io/netty/internal/tcnative/SSL.java
@@ -29,7 +29,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package io.netty.tcnative.jni;
+package io.netty.internal.tcnative;
 
 import java.nio.ByteBuffer;
