+++ openssl-dynamic/src/main/java/io/netty/internal/tcnative/SSLContext.java
@@ -30,7 +30,7 @@
  *  limitations under the License.
  */
 
-package io.netty.tcnative.jni;
+package io.netty.internal.tcnative;
 
 public final class SSLContext {
