+++ pom.xml
@@ -246,7 +246,7 @@
                 <supportedProjectType>jar</supportedProjectType>
               </supportedProjectTypes>
               <instructions>
-                <Export-Package>io.netty.tcnative.jni.*</Export-Package>
+                <Export-Package>io.netty.internal.tcnative.*</Export-Package>
                 <Bundle-NativeCode>${tcnativeManifest}</Bundle-NativeCode>
               </instructions>
             </configuration>
